+++ src/modules/chichi/command.ts
@@ -1,8 +1,8 @@
 import { SlashCommandBuilder } from '@discordjs/builders';
 
 export const command = new SlashCommandBuilder()
-  .setName('gege')
-  .setDescription('List all commands available with GérardBot.');
+  .setName('chichi')
+  .setDescription('List all commands available with Chicanbot.');
 
 command.addBooleanOption((option) => option
 	.setName('show_everyone')
@@ -10,6 +10,6 @@ command.addBooleanOption((option) => option
 );
 
 export const help =
-`List all commands available with GérardBot.
+`List all commands available with Chicanbot.
 By default the list is only visible to the user who executed the command.
 To show the list to a whole channel, set the \`showEveryone\`to true`;