+++ src/modules/index.ts
@@ -1,2 +1,2 @@
 export * as Poll from './poll';
-export * as Gege from './gege';
+export * as Chichi from './chichi';